+++ browser/DEPS
@@ -38,6 +38,7 @@ include_rules += [
   "+brave/components/brave_rewards/core",
   "+brave/components/brave_shields/content/browser",
   "+brave/components/brave_shields/core/common",
+  "+brave/components/brave_shields/core/browser",
   "+brave/components/brave_wallet/browser",
   "+brave/components/brave_wallet/common",
   "+brave/components/containers/buildflags",
