+++ browser/android/preferences/BUILD.gn
@@ -16,7 +16,7 @@ source_set("preferences") {
     "//brave/components/brave_perf_predictor/common",
     "//brave/components/brave_referrals/common",
     "//brave/components/brave_rewards/core",
-    "//brave/components/brave_shields/content/browser",
+    "//brave/components/brave_shields/core/browser",
     "//brave/components/brave_sync:prefs",
     "//brave/components/constants",
     "//chrome/browser:browser_process",
