+++ browser/profiles/BUILD.gn
@@ -12,7 +12,7 @@ source_set("util") {
   deps = [
     "//base",
     "//brave/common",
-    "//brave/components/brave_shields/content/browser",
+    "//brave/components/brave_shields/core/browser",
     "//brave/components/brave_shields/core/common",
     "//brave/components/ntp_background_images/browser",
     "//brave/components/ntp_background_images/common",
