+++ browser/ui/webui/brave_shields/shields_panel_handler.cc
@@ -10,7 +10,7 @@
 
 #include "base/check.h"
 #include "brave/browser/ui/brave_browser_window.h"
-#include "brave/components/brave_shields/content/browser/brave_shields_p3a.h"
+#include "brave/components/brave_shields/core/browser/brave_shields_p3a.h"
 #include "brave/components/constants/pref_names.h"
 #include "chrome/browser/browser_process.h"
 #include "chrome/browser/profiles/profile.h"
