+++ browser/ui/webui/brave_shields/shields_panel_ui.cc
@@ -9,7 +9,7 @@
 
 #include "base/check.h"
 #include "brave/browser/ui/brave_browser_window.h"
-#include "brave/components/brave_shields/content/browser/brave_shields_util.h"
+#include "brave/components/brave_shields/core/browser/brave_shields_utils.h"
 #include "brave/components/brave_shields/core/common/brave_shield_localized_strings.h"
 #include "brave/components/brave_shields/core/common/features.h"
 #include "brave/components/brave_shields/resources/panel/grit/brave_shields_panel_generated_map.h"
