+++ chromium_src/chrome/browser/content_settings/DEPS
@@ -1,5 +1,6 @@
 include_rules = [
   "+brave/components/brave_shields/content/browser",
+  "+brave/components/brave_shields/core/browser",
   "+brave/components/brave_shields/core/common",
   "+brave/components/content_settings/core/browser",
 ]
