+++ chromium_src/chrome/browser/prefs/DEPS
@@ -6,7 +6,7 @@ include_rules = [
   "+brave/components/brave_news/common/p3a_pref_names.h",
   "+brave/components/brave_news/common/pref_names.h",
   "+brave/components/brave_search_conversion/p3a.h",
-  "+brave/components/brave_shields/content/browser/brave_shields_p3a.h",
+  "+brave/components/brave_shields/core/browser/brave_shields_p3a.h",
   "+brave/components/brave_sync/brave_sync_prefs.h",
   "+brave/components/brave_vpn/common/buildflags/buildflags.h",
   "+brave/components/brave_wallet/browser/brave_wallet_prefs.h",
