+++ chromium_src/chrome/browser/ssl/DEPS
@@ -1,3 +1,3 @@
 include_rules = [
-  "+brave/components/brave_shields/content/browser",
+  "+brave/components/brave_shields/core/browser",
 ]
