+++ chromium_src/chrome/browser/ssl/https_upgrades_interceptor.cc
@@ -6,7 +6,7 @@
 #include "chrome/browser/ssl/https_upgrades_interceptor.h"
 
 #include "brave/browser/brave_browser_process.h"
-#include "brave/components/brave_shields/content/browser/brave_shields_util.h"
+#include "brave/components/brave_shields/core/browser/brave_shields_utils.h"
 #include "chrome/browser/content_settings/host_content_settings_map_factory.h"
 #include "net/base/features.h"
 #include "net/base/url_util.h"
