+++ chromium_src/chrome/browser/ssl/https_upgrades_navigation_throttle.cc
@@ -6,7 +6,7 @@
 #include "chrome/browser/ssl/https_upgrades_navigation_throttle.h"
 
 #include "base/time/time.h"
-#include "brave/components/brave_shields/content/browser/brave_shields_util.h"
+#include "brave/components/brave_shields/core/browser/brave_shields_utils.h"
 #include "chrome/browser/content_settings/host_content_settings_map_factory.h"
 #include "chrome/browser/profiles/profile.h"
 #include "components/prefs/pref_service.h"
