+++ components/brave_shields/content/browser/BUILD.gn
@@ -34,8 +34,6 @@ static_library("browser") {
     "blocked_domain_1pes_lifetime.h",
     "brave_farbling_service.cc",
     "brave_farbling_service.h",
-    "brave_shields_p3a.cc",
-    "brave_shields_p3a.h",
     "brave_shields_util.cc",
     "brave_shields_util.h",
     "cookie_list_opt_in_service.cc",
