+++ components/brave_shields/content/browser/domain_block_navigation_throttle.cc
@@ -20,6 +20,7 @@
 #include "brave/components/brave_shields/content/browser/domain_block_controller_client.h"
 #include "brave/components/brave_shields/content/browser/domain_block_page.h"
 #include "brave/components/brave_shields/content/browser/domain_block_tab_storage.h"
+#include "brave/components/brave_shields/core/browser/brave_shields_utils.h"
 #include "brave/components/brave_shields/core/common/features.h"
 #include "brave/components/ephemeral_storage/ephemeral_storage_service.h"
 #include "brave/content/public/browser/devtools/adblock_devtools_instumentation.h"
