+++ components/brave_shields/content/browser/domain_block_navigation_throttle.h
@@ -10,7 +10,7 @@
 
 #include "base/memory/raw_ptr.h"
 #include "base/memory/weak_ptr.h"
-#include "brave/components/brave_shields/content/browser/brave_shields_util.h"
+#include "brave/components/brave_shields/core/browser/brave_shields_utils.h"
 #include "content/public/browser/navigation_throttle.h"
 
 class HostContentSettingsMap;
