+++ components/brave_shields/core/browser/DEPS
@@ -1,5 +1,6 @@
 include_rules = [
   "+components/component_updater",
+  "+components/content_settings/core",
   "+components/prefs",
   "+components/value_store",
   "+crypto",
