+++ components/misc_metrics/BUILD.gn
@@ -31,7 +31,7 @@ static_library("misc_metrics") {
     "//base",
     "//brave/components/brave_new_tab_ui:mojom",
     "//brave/components/brave_rewards/core",
-    "//brave/components/brave_shields/content/browser",
+    "//brave/components/brave_shields/core/browser",
     "//brave/components/p3a_utils",
     "//brave/components/time_period_storage",
     "//components/autofill/core/browser",
